+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.3](https://github.com/charlesgael/tauri-pinia/compare/v1.0.1...v1.0.3) (2023-10-03)
+- New feature to allow non-persistant attributes or stores
+- Bug fix for default values of properties in case not persisted on disk
+
 ### [1.0.2](https://github.com/charlesgael/tauri-pinia/compare/v1.0.1...v1.0.2) (2022-06-16)
 
 ### 1.0.1 (2022-06-14)